+++ src/main/java/com/openggf/game/sonic1/objects/Sonic1JunctionObjectInstance.java
@@ -24,6 +24,7 @@ import com.openggf.level.objects.SolidObjectListener;
 import com.openggf.level.objects.SolidObjectParams;
 import com.openggf.level.objects.SolidObjectProvider;
 import com.openggf.level.objects.SolidRoutineProfile;
+import com.openggf.level.objects.SpawnRewindRecreatable;
 import com.openggf.level.render.PatternSpriteRenderer;
 import com.openggf.sprites.playable.AbstractPlayableSprite;
 import com.openggf.sprites.playable.ObjectControlState;
@@ -570,7 +571,8 @@ public class Sonic1JunctionObjectInstance extends AbstractObjectInstance
      * behind the main junction disc. In the ROM, this is a separate SST entry created
      * by Jun_Main with obRoutine=4 (Jun_Display), which just calls RememberState.
      */
-    static class Sonic1JunctionChildInstance extends AbstractObjectInstance {
+    static class Sonic1JunctionChildInstance extends AbstractObjectInstance
+            implements SpawnRewindRecreatable {
 
         Sonic1JunctionChildInstance(ObjectSpawn parentSpawn) {
             super(parentSpawn, "JunctionChild");